심화과제 복잡한 App.tsx 컴포넌트 분리 완료.
