+++ src/hooks/useEventForm.ts
@@ -2,29 +2,35 @@ import { ChangeEvent, useState } from 'react';
 
 import { Event, RepeatType } from '../types';
 import { getTimeErrorMessage } from '../utils/timeValidation';
+import {useAtom} from "jotai/index";
+import {
+  categoryAtom,
+  dateAtom,
+  descriptionAtom,
+  editingEventAtom,
+  endTimeAtom, isRepeatingAtom,
+  locationAtom, notificationTimeAtom, repeatEndDateAtom, repeatIntervalAtom, repeatTypeAtom,
+  startTimeAtom, timeErrorAtom,
+  titleAtom
+} from "../stores/stores.ts";
 
 type TimeErrorRecord = Record<'startTimeError' | 'endTimeError', string | null>;
 
 export const useEventForm = (initialEvent?: Event) => {
-  const [title, setTitle] = useState(initialEvent?.title || '');
-  const [date, setDate] = useState(initialEvent?.date || '');
-  const [startTime, setStartTime] = useState(initialEvent?.startTime || '');
-  const [endTime, setEndTime] = useState(initialEvent?.endTime || '');
-  const [description, setDescription] = useState(initialEvent?.description || '');
-  const [location, setLocation] = useState(initialEvent?.location || '');
-  const [category, setCategory] = useState(initialEvent?.category || '');
-  const [isRepeating, setIsRepeating] = useState(initialEvent?.repeat.type !== 'none');
-  const [repeatType, setRepeatType] = useState<RepeatType>(initialEvent?.repeat.type || 'none');
-  const [repeatInterval, setRepeatInterval] = useState(initialEvent?.repeat.interval || 1);
-  const [repeatEndDate, setRepeatEndDate] = useState(initialEvent?.repeat.endDate || '');
-  const [notificationTime, setNotificationTime] = useState(initialEvent?.notificationTime || 10);
-
-  const [editingEvent, setEditingEvent] = useState<Event | null>(null);
-
-  const [{ startTimeError, endTimeError }, setTimeError] = useState<TimeErrorRecord>({
-    startTimeError: null,
-    endTimeError: null,
-  });
+  const [title, setTitle] = useAtom(titleAtom);
+  const [date, setDate] = useAtom(dateAtom);
+  const [startTime, setStartTime] = useAtom(startTimeAtom);
+  const [endTime, setEndTime] = useAtom(endTimeAtom);
+  const [description, setDescription] = useAtom(descriptionAtom);
+  const [location, setLocation] = useAtom(locationAtom);
+  const [category, setCategory] = useAtom(categoryAtom);
+  const [isRepeating, setIsRepeating] = useAtom(isRepeatingAtom);
+  const [repeatType, setRepeatType] = useAtom(repeatTypeAtom);
+  const [repeatInterval, setRepeatInterval] = useAtom(repeatIntervalAtom);
+  const [repeatEndDate, setRepeatEndDate] = useAtom(repeatEndDateAtom);
+  const [notificationTime, setNotificationTime] = useAtom(notificationTimeAtom);
+  const [editingEvent, setEditingEvent] = useAtom(editingEventAtom);
+  const [{ startTimeError, endTimeError }, setTimeError] = useAtom(timeErrorAtom);
 
   const handleStartTimeChange = (e: ChangeEvent<HTMLInputElement>) => {
     const newStartTime = e.target.value;
