+++ src/ui/ScheduleView/calendar/index.ts
@@ -1,0 +1,2 @@
+export { default as MonthCalendar } from "./MonthCalendar.tsx";
+export { default as WeekCalendar } from "./WeekCalendar.tsx";
