+++ src/utils/types.ts
@@ -1,0 +1,9 @@
+export interface Notification {
+  id: string;
+  message: string;
+}
+
+export interface TimeErrorRecord {
+  startTimeError: string | null;
+  endTimeError: string | null;
+}
